pyproject.toml
[project]
name = "ai-perf-portfolio"
version = "0.1.0"
description = "Single-GPU AI performance engineering portfolio (benchmarks, RAG, observability, multi-GPU scaffold)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.4",
    "rich>=13.7",
    "typer>=0.12",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "prometheus-client>=0.17",
    "structlog>=24.1",
    "opentelemetry-api>=1.21",
    "opentelemetry-sdk>=1.21",
    "opentelemetry-exporter-otlp>=1.21",
    "httpx>=0.25",
    "fastapi>=0.105",
    "uvicorn>=0.24",
    "pytest>=7.4",
    "pytest-benchmark>=4.0",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5",
    "transformers>=4.34",
    "accelerate>=0.24",
    "safetensors>=0.4",
]

[project.optional-dependencies]
gpu = [
    # To be installed when GPU is present; keep commented or install manually
    # "torch==2.1.0+cu121",
    # "xformers",
    # "flash-attn",
    # "faiss-gpu",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
